Allow for anchors in news and redirect permalinks.
A landing page design is not the right solution for posting news.
The main problem resulted in the lack of a good redirection schema
for permalinks.

With this change we can set a variable that will be able to redirect
to the middle of the landing page corresponding to the news.